Refactor cached code lists to a custom hook to re-use the code between regular Work and BatchEdit components
